+++ CHANGELOG.md
@@ -9,6 +9,9 @@
 - Add `Nimrag.steps_weekly`
 - Add `Nimrag.sleep_daily`
 - Add `Nimrag.user_settings`
+- Rename struct `Nimrag.Api.Activity` to `Nimrag.Api.ActivityList` (data struct for elements of activities list)
+- Add `Nimrag.activity`
+- Add `Nimrag.activity_details`
 
 ## 0.1.0 (2024-03-29)